fix(core): move overlay to ensure old behaviour of the overlay being placed over all elements
